agent-framework/requirements.txt
deepagents>=0.5.0
langchain>=1.0.0
langchain-openai>=1.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pyyaml>=6.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
sse-starlette>=2.0.0
